+++ src/Exception/EntryDoesNotExistException.php
@@ -6,4 +6,15 @@ use DomainException;
 
 final class EntryDoesNotExistException extends DomainException implements Exception
 {
+    use ExceptionFactory;
+
+    /**
+     * @param string key
+     *
+     * @return self
+     */
+    public static function fromKey($key)
+    {
+        return self::create('No entry found for "%s".', $key);
+    }
 }