+++ src/Exception/FileNotFoundException.php
@@ -2,6 +2,19 @@
 
 namespace TomPHP\ConfigServiceProvider\Exception;
 
-class FileNotFoundException extends \RuntimeException implements Exception
+use LogicException;
+
+final class FileNotFoundException extends LogicException implements Exception
 {
+    use ExceptionFactory;
+
+    /**
+     * @param string $filename
+     *
+     * @return self
+     */
+    public static function fromFileName($filename)
+    {
+        return self::create('"%s" does not exist', $filename);
+    }
 }