+++ src/Exception/NotClassDefinitionException.php
@@ -2,6 +2,22 @@
 
 namespace TomPHP\ConfigServiceProvider\Exception;
 
-final class NotClassDefinitionException extends \RuntimeException implements Exception
+use LogicException;
+
+final class NotClassDefinitionException extends LogicException implements Exception
 {
+    use ExceptionFactory;
+
+    /**
+     * @param string $name
+     *
+     * @return self
+     */
+    public static function fromServiceName($name)
+    {
+        return self::create(
+            'Service configuration for "%s" did not create a class definition.',
+            $name
+        );
+    }
 }