+++ src/Exception/ReadOnlyException.php
@@ -2,6 +2,19 @@
 
 namespace TomPHP\ConfigServiceProvider\Exception;
 
-final class ReadOnlyException extends \LogicException
+use LogicException;
+
+final class ReadOnlyException extends LogicException implements Exception
 {
+    use ExceptionFactory;
+
+    /**
+     * @param string $name
+     *
+     * @return self
+     */
+    public static function fromClassName($name)
+    {
+        return self::create('"%s" is read only.', $name);
+    }
 }