+++ tests/Exception/InvalidConfigExceptionTest.php
@@ -15,4 +15,24 @@ final class InvalidConfigExceptionTest extends PHPUnit_Framework_TestCase
         );
     }
 
+    public function testItIsALogicException()
+    {
+        $this->assertInstanceOf('LogicException', new InvalidConfigException());
+    }
+
+    public function testItCanBeCreatedFromTheFileName()
+    {
+        $this->assertEquals(
+            '"example.cfg" does not return a PHP array.',
+            InvalidConfigException::fromPHPFileError('example.cfg')->getMessage()
+        );
+    }
+
+    public function testItCanBeCreatedWithAJSONFileError()
+    {
+        $this->assertEquals(
+            'Invalid JSON in "example.json": JSON Error Message',
+            InvalidConfigException::fromJSONFileError('example.json', 'JSON Error Message')->getMessage()
+        );
+    }
 }