+++ tests/Exception/NotClassDefinitionExceptionTest.php
@@ -14,4 +14,17 @@ final class NotClassDefinitionExceptionTest extends PHPUnit_Framework_TestCase
             new NotClassDefinitionException()
         );
     }
+
+    public function testItIsALogicException()
+    {
+        $this->assertInstanceOf('LogicException', new NotClassDefinitionException());
+    }
+
+    public function testItCanBeCreatedFromThePatterns()
+    {
+        $this->assertEquals(
+            'Service configuration for "example-service" did not create a class definition.',
+            NotClassDefinitionException::fromServiceName('example-service')->getMessage()
+        );
+    }
 }